requirements.txt
python-telegram-bot==13.12
psycopg2-binary==2.9.6